requirements.txt
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
